contests/mine_problem/3nd place solution/code/requirements.txt
pandas~=1.2.1
numpy~=1.20.1
torch~=1.9.0+cu111
torchvision~=0.10.0+cu111
matplotlib~=3.4.2
Pillow~=8.1.0
bbaug~=0.4.2
scikit-learn~=0.24.2
plotly~=4.14.3
pycocotools~=2.0